Refactor: Extract TechStackSelectDialog into individual file
