+++ src/components/Header/AddProjectDialog/AddProjectDialog.scss
@@ -9,9 +9,11 @@
 }
 
 .add-project-dialog-title {
+  width: 100%;
   margin-top: 16px;
   font-size: 24px;
   font-weight: bold;
+  text-align: center;
 }
 
 .add-project-dialog-subtitle {
