+++ src/components/PortfolioPage/UsingTechStack/UsingTechStack.jsx
@@ -4,43 +4,8 @@ import AddCircleOutlineIcon from '@mui/icons-material/AddCircleOutline';
 import Title from '../../Title/Title';
 import './UsingTechStack.scss';
 import Conditional from '../../Conditional/Conditional';
-import {
-  Dialog,
-  List,
-  ListItem,
-  ListItemIcon,
-  ListItemText,
-} from '@mui/material';
 import { useState } from 'react';
-
-function TechStackSelectDialog({ onClose, open, data }) {
-  const handleClose = (stack) => {
-    onClose(stack);
-  };
-
-  return (
-    <Dialog onClose={handleClose} open={open} fullWidth maxWidth="md">
-      <List>
-        {data.map((v) => (
-          <ListItem
-            key={`tech-stack-select-dialog-${v.id}`}
-            disablePadding
-            onClick={() => handleClose(v)}
-          >
-            <ListItemIcon>
-              <img
-                className="using-tech-stack-dialog-icon"
-                src={v.iconUrl}
-                alt="icon"
-              />
-            </ListItemIcon>
-            <ListItemText primary={v.name} />
-          </ListItem>
-        ))}
-      </List>
-    </Dialog>
-  );
-}
+import TechStackSelectDialog from '../../TechStackSelectDialog/TechStackSelectDialog';
 
 function UsingTechStackUnit({ data, isEditMode }) {
   const stacks = useSelector((state) => state.techStack.allTechStack.data);
