+++ src/components/TechStackSelectDialog/TechStackSelectDialog.scss
@@ -1,0 +1,7 @@
+.tech-stack-select-dialog-icon {
+  border-radius: 18px;
+  width: 36px;
+  height: 36px;
+  object-fit: cover;
+  margin: 8px;
+}
